It begins! Restructuring flight control functions to be more modular.
This is necessary for properly implementing multirotors (or VTOLs)
