+++ Autopilot/AttitudeManager/StateMachine.h
@@ -8,10 +8,6 @@
 #ifndef STATEMACHINE_H
 #define	STATEMACHINE_H
 
-#ifdef	__cplusplus
-extern "C" {
-#endif
-
 #define STATEMACHINE_MAINLOOP 0
 #define STATEMACHINE_IMU 1
 
@@ -24,15 +20,8 @@ extern "C" {
 #include "VN100.h"
 #include "cameraManager.h"
 
-    void StateMachine(char entryLocation);
-    void highLevelControl();
-    void lowLevelControl();
-    void forceStateMachineUpdate();
-    void killPlane(char action);
-
-#ifdef	__cplusplus
-}
-#endif
+void StateMachine(char entryLocation);
+void forceStateMachineUpdate();
+void killPlane(char action);
 
 #endif	/* STATEMACHINE_H */
-
