+++ Autopilot/AttitudeManager/nbproject/configurations.xml
@@ -7,6 +7,7 @@
       <logicalFolder name="f9" displayName="Attitude Management" projectFiles="true">
         <itemPath>AttitudeManager.h</itemPath>
         <itemPath>OrientationControl.h</itemPath>
+        <itemPath>StateMachine.h</itemPath>
       </logicalFolder>
       <logicalFolder name="f6" displayName="Common" projectFiles="true">
         <itemPath>../Common/Common.h</itemPath>
@@ -50,7 +51,7 @@
       </logicalFolder>
       <logicalFolder name="f11" displayName="Vehicles" projectFiles="true">
         <itemPath>Anaconda.h</itemPath>
-        <itemPath>DisplayQuad.h</itemPath>
+        <itemPath>Multirotor.h</itemPath>
       </logicalFolder>
       <itemPath>main.h</itemPath>
     </logicalFolder>
@@ -65,7 +66,6 @@
         <itemPath>AttitudeManager.c</itemPath>
         <itemPath>OrientationControl.c</itemPath>
         <itemPath>StateMachine.c</itemPath>
-        <itemPath>StateMachine.h</itemPath>
       </logicalFolder>
       <logicalFolder name="f7" displayName="Common" projectFiles="true">
         <itemPath>../Common/Common.c</itemPath>
@@ -109,7 +109,7 @@
       </logicalFolder>
       <logicalFolder name="f11" displayName="Vehicles" projectFiles="true">
         <itemPath>Anaconda.c</itemPath>
-        <itemPath>DisplayQuad.c</itemPath>
+        <itemPath>Multirotor.c</itemPath>
       </logicalFolder>
       <itemPath>main.c</itemPath>
     </logicalFolder>
